refactor serialization logic to library crate
