--- example/Cargo.toml
@@ -1,7 +1,0 @@
-[workspace]
-members = ["script"]
-exclude = ["solana/program"]
-resolver = "2"
-
-[workspace.dependencies]
-groth16-solana = { path = "../verifier" }
