+++ example/script/Cargo.toml
@@ -5,11 +5,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.0", features = ["derive", "env"] }
-hex = "0.4.3"
-num-bigint = "0.4.6"
-num-traits = "0.2.19"
-groth16-solana.workspace = true
+clap = { version = "4.0", features = ["derive"] }
+groth16-solana = { path = "../../verifier", features = ["sp1-serialize"] }
 sp1-sdk = "2.0.0"
 strum = "0.25"
 strum_macros = "0.25"
