+++ verifier/Cargo.toml
@@ -16,12 +16,15 @@ ark-serialize = "0.4.2"
 ark-ff = "0.4.2"
 borsh = "1.5.1"
 solana-bn254 = { git = "https://github.com/anza-xyz/agave" }
-# solana-program = { git = "https://github.com/anza-xyz/agave" }
-# groth16-solana = { git = "https://github.com/sp1-patches/groth16-solana", branch = "patch-v0.0.3" }
-groth16-solana = { path = "../../groth16-solana-patch" }
+groth16-solana = { git = "https://github.com/sp1-patches/groth16-solana", branch = "patch-v0.0.3" }
 thiserror = "1.0.63"
 
-# sp1-sdk = { version = "2.0.0", default-features = false }
-hex = "0.4.3"
-num-bigint = "0.4.6"
-num-traits = "0.2.19"
+# Optional dependencies for SP1 serialization
+sp1-sdk = { version = "2.0.0", default-features = false, optional = true }
+hex = { version = "0.4.3", optional = true }
+num-bigint = { version = "0.4.6", optional = true }
+num-traits = { version = "0.2.19", optional = true }
+
+[features]
+default = []
+sp1-serialize = ["dep:sp1-sdk", "dep:hex", "dep:num-bigint", "dep:num-traits"]
